requirements.txt
flask==3.0.2
azure-cosmos==4.2.0
azure.storage.blob==12.19.1